+++ react/typings/simple-element-resize-detector.d.ts
@@ -1,0 +1,6 @@
+declare module 'simple-element-resize-detector' {
+  export default function(
+    el: Element,
+    handler: (el: Element) => void
+  ): HTMLIFrameElement
+}