+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `HighlightOverlay` now has a new look: No more background color and dashed border, also added a label to show which block is selected. It stays visible when the user is editing a content.
+
 ## [4.12.7] - 2019-10-08
 
 ### Fixed